added changes suggested by ethan
