+++ driving-day-app-frontend/src/api/api.ts
@@ -1,7 +1,9 @@
 import axios from "axios";
 import { AxiosError } from "axios";
 import exp from "constants";
-import { DataCategory, ResponseValue, PackingListEntry  } from "../utils/DataTypes";
+import { DataCategory} from "../types/Chart";
+import { ResponseValue } from "../types/Feedback";
+import { PackingListEntry } from "../types/PackingList";
 
 export const api = axios.create({
   baseURL: `${process.env.REACT_APP_BACKEND_URL}/api/`,
@@ -123,6 +125,7 @@ export const postFeedback = async (feedbackData: {
 };
 
 
+
 /**
  *
  * @param path: Corresponds to the relative backend path to which the GET request is sent
@@ -135,6 +138,28 @@ export const getRequest = async (
 ) => {
   try {
     const response = await api.get(`${path}?${searchParams.toString()}`);
+    return response;
+  } catch (error) {
+    console.error(error);
+    const axiosError = error as AxiosError;
+    return { status: axiosError.status, data: undefined };
+  }
+};
+
+export const putRequest = async (path: string, data: any) => {
+  try {
+    const response = await api.put(path, data);
+    return response;
+  } catch (error) {
+    console.error(error);
+    const axiosError = error as AxiosError;
+    return { status: axiosError.status, data: undefined };
+  }
+};
+
+export const deleteRequest = async (path: string) => {
+  try {
+    const response = await api.delete(path);
     return response;
   } catch (error) {
     console.error(error);
@@ -346,27 +371,8 @@ export const updateIssue = async (
   }
 };
 
-export const updateDrivingDay = async (
-  drivingDayId: string,
-  drivingDayData: {
-    title?: string;
-    date?: string;
-    description?: string;
-    drivers?: string[];
-    packingLists?: PackingListEntry[];
-    feedback?: number[];
-    issues?: number[];
-  }
-) => {
-  const path = `update-driving-day/${drivingDayId}/`;
-  try {
-    const response = await api.put(path, drivingDayData);
-    return response;
-  } catch (error) {
-    console.error(error);
-    const axiosError = error as AxiosError;
-    return { status: axiosError.status };
-  }
+export const updateDrivingDay = async (drivingDayId: string, drivingDayData: any) => {
+  return await putRequest(`update-driving-day/${drivingDayId}/`, drivingDayData);
 };
 
 export const deleteIssue = async (issueId: string) => {
@@ -382,20 +388,11 @@ export const deleteIssue = async (issueId: string) => {
 };
 
 export const deleteDrivingDay = async (drivingDayId: string) => {
-
-  const path = `delete-driving-day/${drivingDayId}/`;
-  try {
-    const response = await api.delete(path);
-    return response;
-  } catch (error) {
-    console.error(error);
-    const axiosError = error as AxiosError;
-    return { status: axiosError.status };
-  }
+  return await deleteRequest(`delete-driving-day/${drivingDayId}/`);
 };
 
 export const getAllPackingLists = async () => {
-  const path = "all-packing-lists";
+  const path = "get-all-packing-lists";
   return await getRequest(path, new URLSearchParams());
 };
