+++ driving-day-app-frontend/src/components/contexts/ChartContext.tsx
@@ -1,4 +1,4 @@
-import { StandardChartProps } from '../../utils/DataTypes';
+import { StandardChartProps } from '../../types/Chart';
 import { createContext } from 'react';
 
 interface ChartContextType {
