+++ driving-day-app-frontend/src/components/issues-components/AddIssueModal.tsx
@@ -2,7 +2,7 @@ import React, { useContext, useEffect, useState } from "react";
 import Modal from "../shared/Modal";
 import { postIssue, postS3Image } from "../../api/api";
 import { availableSubsystems, priorityLevels, statusOptions } from "../../constants/IssuesConstants";
-import { Issue } from "../../utils/DataTypes";
+import { Issue } from "../../types/Issues";
 import AppDataContext from '../contexts/AppDataContext';
 
 interface AddIssueModalProps {
