+++ driving-day-app-frontend/src/types/PackingList.ts
@@ -1,0 +1,24 @@
+export interface PackingListEntry {
+    packingListId: string;
+    checkedItems: number[];
+}
+
+export interface PackingList {
+    id: string;
+    name: string;
+    description: string;
+    items: string[];
+    category?: PackingCategory;
+    order?: number;
+}
+
+export interface PackingList {
+    id: string;
+    name: string;
+    description: string;
+    items: string[];
+    category?: PackingCategory;
+    order?: number;
+}
+
+export type PackingCategory = "Standard" | "Subsystems";
